+++ src/__tests__/handler/index.test.ts
@@ -1,7 +1,8 @@
 import { Neovim } from '@chemzqm/neovim'
-import { Disposable } from 'vscode-languageserver-protocol'
+import { Disposable, SymbolKind } from 'vscode-languageserver-protocol'
 import commands from '../../commands'
 import Handler from '../../handler/index'
+import { handleError, toDocumentation } from '../../handler/util'
 import { ProviderName } from '../../languages'
 import { disposeAll } from '../../util'
 import helper from '../helper'
@@ -29,10 +30,36 @@ afterEach(async () => {
 })
 
 describe('Handler', () => {
+  describe('util', () => {
+    it('should handleError', () => {
+      handleError(new Error('error'))
+    })
+
+    it('should to documentation', () => {
+      expect(toDocumentation('doc')).toEqual({ content: 'doc', filetype: 'txt' })
+      expect(toDocumentation({ kind: 'markdown', value: 'doc' })).toEqual({ content: 'doc', filetype: 'markdown' })
+    })
+  })
+
   describe('hasProvider', () => {
     it('should check provider for document', async () => {
       let res = await helper.doAction('hasProvider', 'definition')
       expect(res).toBe(false)
+      await nvim.command(`edit +setl\\ buftype=nofile foo`)
+      res = await handler.hasProvider('formatOnType')
+      expect(res).toBe(false)
+    })
+  })
+
+  describe('getIcon', () => {
+    it('should get icon', () => {
+      helper.updateConfiguration('suggest.completionItemKindLabels', {
+        default: 'd'
+      })
+      let res = handler.getIcon(SymbolKind.Array)
+      expect(res).toBeDefined()
+      res = handler.getIcon('a' as any)
+      expect(res.text).toBe('d')
     })
   })